// src/types/player.type.ts
export interface IPlayerRank {
  team_name: string;
  player_name: string;
  season: string;
  team_id: string;
  player_id: string;
  game_count: string;
  game_start: string;
  minute_play: string;
  pts: string;
  fga: string;
  fgm: string;
  fgp: string;
  paint_f2ga: string;
  paint_f2gm: string;
  paint_f2gp: string;
  f3ga: string;
  f3gm: string;
  f3gp: string;
  fta: string;
  ftm: string;
  ftp: string;
  orb: string;
  drb: string;
  trb: string;
  ast: string;
  stl: string;
  blk: string;
  tov: string;
  pf: string;
  fb: string;
  pot: string;
  f2ga: string;
  f2gm: string;
  f2gp: string;
  }